Download all translations of the initial set of 50 popular titles
... as top-level topics under the African Storybook channel, scraped from the "translations and adaptations" sidebar menu.

This produces 80 languages with the initial set of 50 titles. The most populated
languages are 80 English titles, 54 Swahili, 43 Zulu, 28 French, 24 Portuguese ...

I've updated the demo server that @ivanistheone set up with the run from this PR.

Take a look ... there's a lot of languages, and most languages don't have many
content items yet, with just those 50 initial items. So we have a long sparse
tail. I'll do a full scrape of the entire index of ~4000 items and see how this
distribution looks! Here's the PR for now, feel free to review @ivanistheone